start of refactoring to use fixture for asp.net core tests
